fix: 플레이리스트핀 삭제 로직 수정
