+++ src/main/java/sws/songpin/domain/pin/service/PinService.java
@@ -3,7 +3,6 @@ package sws.songpin.domain.pin.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import sws.songpin.domain.bookmark.entity.Bookmark;
 import sws.songpin.domain.genre.entity.Genre;
 import sws.songpin.domain.genre.entity.GenreName;
 import sws.songpin.domain.genre.service.GenreService;
@@ -13,14 +12,14 @@ import sws.songpin.domain.pin.dto.request.PinAddRequestDto;
 import sws.songpin.domain.pin.dto.request.PinUpdateRequestDto;
 import sws.songpin.domain.pin.entity.Pin;
 import sws.songpin.domain.pin.repository.PinRepository;
-import sws.songpin.domain.place.dto.request.PlaceAddRequestDto;
 import sws.songpin.domain.place.entity.Place;
 import sws.songpin.domain.place.service.PlaceService;
-import sws.songpin.domain.playlist.service.PlaylistService;
+import sws.songpin.domain.playlist.entity.Playlist;
+import sws.songpin.domain.playlist.repository.PlaylistRepository;
 import sws.songpin.domain.playlistpin.entity.PlaylistPin;
-import sws.songpin.domain.song.dto.request.SongAddRequestDto;
+import sws.songpin.domain.playlistpin.repository.PlaylistPinRepository;
+import sws.songpin.domain.playlistpin.service.PlaylistPinService;
 import sws.songpin.domain.song.dto.response.SongDetailsPinDto;
-import sws.songpin.domain.song.dto.response.SongDetailsResponseDto;
 import sws.songpin.domain.song.entity.Song;
 import sws.songpin.domain.song.repository.SongRepository;
 import sws.songpin.domain.song.service.SongService;
@@ -37,7 +36,8 @@ import java.util.stream.Collectors;
 public class PinService {
     private final PinRepository pinRepository;
     private final SongRepository songRepository;
-    private final PlaylistService playlistService;
+    private final PlaylistPinRepository playlistPinRepository;
+    private final PlaylistPinService playlistPinService;
     private final MemberService memberService;
     private final SongService songService;
     private final PlaceService placeService;
@@ -81,11 +81,12 @@ public class PinService {
     // 음악 핀 삭제
     public void deletePin(Long pinId) {
         Pin pin = validatePinCreator(pinId);
-        List<PlaylistPin> playlistPins = playlistService.getPlaylistPinsByPin(pin);
+        List<PlaylistPin> playlistPins = playlistPinService.getPlaylistPinsByPin(pin);
         for (PlaylistPin playlistPin : playlistPins) {
-            playlistService.removePlaylistPin(playlistPin);
+            Playlist playlist = playlistPin.getPlaylist();
+            playlist.removePlaylistPin(playlistPin);
         }
-        pinRepository.delete(pin);
+        playlistPinRepository.deleteAll(playlistPins);
     }
 
     private void updateSongAvgGenreName(Song song) {
