+++ src/main/java/sws/songpin/domain/playlist/service/PlaylistService.java
@@ -9,7 +9,6 @@ import sws.songpin.domain.bookmark.repository.BookmarkRepository;
 import sws.songpin.domain.member.entity.Member;
 import sws.songpin.domain.member.service.MemberService;
 import sws.songpin.domain.model.Visibility;
-import sws.songpin.domain.pin.entity.Pin;
 import sws.songpin.domain.playlist.dto.response.*;
 import sws.songpin.domain.playlist.dto.request.PlaylistAddRequestDto;
 import sws.songpin.domain.playlist.dto.request.PlaylistUpdateRequestDto;
@@ -122,9 +121,11 @@ public class PlaylistService {
                 updatedPins.add(currentPin);
             }
         }
-        for (PlaylistPin playlistPin : pinsToDelete) {
-            removePlaylistPin(playlistPin);
+        for (PlaylistPin pin : pinsToDelete) {
+            playlist.removePlaylistPin(pin);
         }
+        playlistPinRepository.deleteAll(pinsToDelete);
+
         playlistPinRepository.saveAll(updatedPins);
     }
 
@@ -194,12 +195,4 @@ public class PlaylistService {
                 .collect(Collectors.toList());
     }
 
-    @Transactional(readOnly = true)
-    public List<PlaylistPin> getPlaylistPinsByPin(Pin pin){
-        List<PlaylistPin> playlistPins = playlistPinRepository.findAllByPin(pin);
-        if (playlistPins.isEmpty()) {
-            throw new CustomException((ErrorCode.PLAYLIST_PIN_NOT_FOUND));
-        }
-        return playlistPins;
-    }
 }
