+++ src/main/java/sws/songpin/domain/playlistpin/service/PlaylistPinService.java
@@ -10,8 +10,11 @@ import sws.songpin.domain.playlist.entity.Playlist;
 import sws.songpin.domain.playlist.repository.PlaylistRepository;
 import sws.songpin.domain.playlist.service.PlaylistService;
 import sws.songpin.domain.playlistpin.entity.PlaylistPin;
+import sws.songpin.domain.playlistpin.repository.PlaylistPinRepository;
 import sws.songpin.global.exception.CustomException;
 import sws.songpin.global.exception.ErrorCode;
+
+import java.util.List;
 
 @Service
 @Transactional
@@ -20,6 +23,7 @@ public class PlaylistPinService {
     private final PlaylistService playlistService;
     private final PinService pinService;
     private final PlaylistRepository playlistRepository;
+    private final PlaylistPinRepository playlistPinRepository;
 
     // 플레이리스트에 핀 추가
     public void addPlaylistPin(PlaylistPinAddRequestDto requestDto) {
@@ -48,6 +52,15 @@ public class PlaylistPinService {
                 .playlist(playlist)
                 .pin(pin)
                 .build();
+    }
+
+    @Transactional(readOnly = true)
+    public List<PlaylistPin> getPlaylistPinsByPin(Pin pin){
+        List<PlaylistPin> playlistPins = playlistPinRepository.findAllByPin(pin);
+        if (playlistPins.isEmpty()) {
+            throw new CustomException((ErrorCode.PLAYLIST_PIN_NOT_FOUND));
+        }
+        return playlistPins;
     }
 
 }
